refactoring datePicker
